Fixs a bug in the logic for the compare module.
There was  bug int he compare module that returned true for
any time that either of the parameters passed in was undefined.
That was an edge case, but probably caused some issues.

There was also a place in the .off module that was incorrectly
using compare by passing it the result of a comparison and thus
always returning true.